+++ src/Pages/Params/Components/index.js
@@ -1,0 +1,2 @@
+export { default as CosmosParamsList } from './CosmosParamsList';
+export { default as ProvParamsList } from './ProvParamsList';