+++ src/consts/urls.js
@@ -60,13 +60,14 @@ export const SOCIAL_SLACK_URL =
   'https://join.slack.com/t/provenanceio/shared_invite/zt-vysymaqb-VqgW3frXoNNXNlyOiP7mog';
 export const SOCIAL_PROVENANCE_URL = 'https://provenance.io/';
 // --Stats
-export const ID_INFO_URL = `${BASE_URL}/chain/id`;
-export const UPGRADE_INFO_URL = `${BASE_URL}/chain/upgrades`;
-export const FEE_STATS_URL = `${BASE_URL}/gas/fees/statistics`;
-export const GAS_STATS_URL = `${BASE_URL}/gas/stats`;
-export const GAS_VOL_URL = `${BASE_URL}/gas/volume`;
-export const SPOTLIGHT_STATS_URL = `${BASE_URL}/spotlight`;
-export const TOKEN_STATS_URL = `${BASE_URL}/token/stats`;
+export const NETWORK_ID_URL = `${BASE_URL}/chain/id`;
+export const NETWORK_UPGRADES_URL = `${BASE_URL}/chain/upgrades`;
+export const NETWORK_FEES_URL = `${BASE_URL}/gas/fees/statistics`;
+export const NETWORK_GAS_STATS_URL = `${BASE_URL}/gas/stats`;
+export const NETWORK_GAS_VOL_URL = `${BASE_URL}/gas/volume`;
+export const NETWORK_PARAMS_URL = `${BASE_URL}/params`;
+export const NETWORK_SPOTLIGHT_URL = `${BASE_URL}/spotlight`;
+export const NETWORK_TOKEN_URL = `${BASE_URL}/token/stats`;
 // -- Txs
 export const TX_INFO_URL = `${BASE_URL}/txs`;
 export const TXS_RECENT_URL = `${BASE_URL}/txs/recent`;