+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Add validator icons #248
 - Add IBC UI #130
 - Migrate orderbook to gecko API #255
+- Add chain params to network dropdown #144
 
 ## 2.5.0